requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
